Added casset support and refactored compiling less into new class and added new option to not keep the dir of the less file
